% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method for linreg objects}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{additional arguments}
}
\value{
Named vector of coefficients
}
\description{
Coefficients method for linreg objects
}
