% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Class using QR Decomposition}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object specifying the model}

\item{data}{A data frame containing the variables}
}
\value{
An object of class linreg containing regression results
}
\description{
A class for performing linear regression analysis using QR decomposition
}
