% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Pred method for linreg objects}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{additional arguments}
}
\value{
Vector of predicted values
}
\description{
Pred method for linreg objects
}
