% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Prediction Method for Linear Regression Model}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of class 'linreg' containing the fitted regression model}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric vector containing the predicted (fitted) values
}
\description{
This function extracts predicted values from a fitted linear
regression model of class 'linreg'. These are the fitted values based on
the model coefficients and input data.
}
\examples{
# Fit a linear regression model
model <- linreg(mpg ~ wt + hp, data = mtcars)

# Get predictions
predictions <- pred(model)

}
