% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{qr.linreg}
\alias{qr.linreg}
\title{QR Decomposition information}
\usage{
\method{qr}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{additional arguments}
}
\value{
QR decomposition object
}
\description{
Returns information about the QR decomposition used in the model
}
