% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract residuals from linreg objects}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Residuals from the linear regression model
}
\description{
Extract residuals from linreg objects
}
