% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid}
\alias{resid}
\title{Extract residuals from a model}
\usage{
resid(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments}
}
\value{
Model residuals
}
\description{
Extract residuals from a model
}
