% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals method for linreg objects}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{additional arguments}
}
\value{
Vector of residuals
}
\description{
Residuals method for linreg objects
}
