% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method for linreg objects}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{additional arguments}
}
\description{
Summary method for linreg objects
}
