% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{vcov.linreg}
\alias{vcov.linreg}
\title{Variance-Covariance Matrix method for linreg objects}
\usage{
\method{vcov}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{additional arguments}
}
\value{
Variance-covariance matrix
}
\description{
Returns the variance-covariance matrix of the coefficients using QR decomposition
}
